Fix #8: Asynchronous low-level Telegram bot API implementation

Closes #8

See merge request madhead/telek!12

+++ model/src/commonMain/kotlin/by/dev/madhead/telek/model/ResponseParameters.kt
@@ -11,7 +11,7 @@ import kotlinx.serialization.Serializable
 @TelegramBotAPIType(type = "responseparameters")
 data class ResponseParameters(
     @SerialName("migrate_to_chat_id")
-    val migrateToChatId: Int? = null,
+    val migrateToChatId: Long? = null,
 
     @SerialName("retry_after")
     val retryAfter: Int? = null
